added parallelization, changed training loop to track overfitting

+++ src/loss.rs
@@ -2,13 +2,13 @@ use ndarray::{Array2, Axis};
 
 /// Computes Softmax Cross-Entropy Loss and the gradient w.r.t the pre-softmax logits.
 pub fn cross_entropy_with_logits(
-    logits: &Array2<f64>,
-    targets: &Array2<f64>,
-) -> (f64, Array2<f64>) {
-    let batch_len = logits.nrows() as f64;
+    logits: &Array2<f32>,
+    targets: &Array2<f32>,
+) -> (f32, Array2<f32>) {
+    let batch_len = logits.nrows() as f32;
 
     let max_per_row = logits.map_axis(Axis(1), |row| {
-        row.iter().cloned().fold(f64::NEG_INFINITY, f64::max)
+        row.iter().cloned().fold(f32::NEG_INFINITY, f32::max)
     });
 
     let shifted = logits - &max_per_row.insert_axis(Axis(1));
